requirements.txt
# Healthcare IDP System - Production Dependencies
# Intelligent Document Processing for Group Benefits

# ═══════════════════════════════════════════════════════════════════════════════
# AWS & Cloud
# ═══════════════════════════════════════════════════════════════════════════════
boto3==1.34.18
botocore>=1.34.0

# ═══════════════════════════════════════════════════════════════════════════════
# LLM & AI Frameworks
# ═══════════════════════════════════════════════════════════════════════════════
langchain==0.1.0
langchain-community==0.0.13
langchain-core>=0.1.0
# openai>=1.0.0  # Optional for OpenAI fallback

# ═══════════════════════════════════════════════════════════════════════════════
# NLP & Machine Learning
# ═══════════════════════════════════════════════════════════════════════════════
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
# sentence-transformers>=2.2.0  # For embeddings

# ═══════════════════════════════════════════════════════════════════════════════
# OCR & Document Processing
# ═══════════════════════════════════════════════════════════════════════════════
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# ═══════════════════════════════════════════════════════════════════════════════
# API & Web Framework
# ═══════════════════════════════════════════════════════════════════════════════
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
aiofiles>=23.0.0

# ═══════════════════════════════════════════════════════════════════════════════
# Data Storage & Caching
# ═══════════════════════════════════════════════════════════════════════════════
sqlalchemy==2.0.25
redis==5.0.1
# pymongo>=4.0.0  # Optional for MongoDB

# ═══════════════════════════════════════════════════════════════════════════════
# Configuration & Environment
# ═══════════════════════════════════════════════════════════════════════════════
python-dotenv==1.0.0
pyyaml==6.0.1
python-json-logger==2.0.7

# ═══════════════════════════════════════════════════════════════════════════════
# Testing & Quality
# ═══════════════════════════════════════════════════════════════════════════════
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
pytest-mock>=3.0.0

# ═══════════════════════════════════════════════════════════════════════════════
# Monitoring & Observability
# ═══════════════════════════════════════════════════════════════════════════════
prometheus-client>=0.17.0
# opentelemetry-api>=1.0.0  # Optional for tracing
